requirements.txt
aiogram==3.1.1
asyncpg==0.27.0
python-dotenv==1.0.0

logging
aiosqlite 